+++ dev/extra-ml-requirements.txt
@@ -40,7 +40,7 @@ onnxruntime
 # mleap format via ``mlflow.spark.log_model``, ``mlflow.spark.save_model``
 mleap
 # Required by mlflow.spark
-pyspark==2.4.0
+pyspark
 # Required by mlflow.shap
 shap
 # Remove `ipython` once this issue: https://github.com/slundberg/shap/pull/1749 has been addressed
