+++ tests/spark/test_spark_model_export.py
@@ -19,7 +19,7 @@ import yaml
 import mlflow
 import mlflow.pyfunc.scoring_server as pyfunc_scoring_server
 import mlflow.tracking
-from mlflow import pyfunc, mleap
+from mlflow import pyfunc
 from mlflow import spark as sparkm
 from mlflow.exceptions import MlflowException
 from mlflow.models import Model, infer_signature
@@ -269,32 +269,16 @@ def test_transformer_model_export(spark_model_transformer, model_path, spark_cus
 @pytest.mark.large
 def test_model_deployment(spark_model_iris, model_path, spark_custom_env):
     sparkm.save_model(
-        spark_model_iris.model,
-        path=model_path,
-        conda_env=spark_custom_env,
-        # Test both spark ml and mleap
-        sample_input=spark_model_iris.spark_df,
+        spark_model_iris.model, path=model_path, conda_env=spark_custom_env,
     )
-
-    # 1. score and compare pyfunc deployed in Sagemaker docker container
-    scoring_response_1 = score_model_in_sagemaker_docker_container(
+    scoring_response = score_model_in_sagemaker_docker_container(
         model_uri=model_path,
         data=spark_model_iris.pandas_df,
         content_type=pyfunc_scoring_server.CONTENT_TYPE_JSON_SPLIT_ORIENTED,
         flavor=mlflow.pyfunc.FLAVOR_NAME,
     )
     np.testing.assert_array_almost_equal(
-        spark_model_iris.predictions, np.array(json.loads(scoring_response_1.content)), decimal=4
-    )
-    # 2. score and compare mleap deployed in Sagemaker docker container
-    scoring_response_2 = score_model_in_sagemaker_docker_container(
-        model_uri=model_path,
-        data=spark_model_iris.pandas_df.to_json(orient="split"),
-        content_type=pyfunc_scoring_server.CONTENT_TYPE_JSON,
-        flavor=mlflow.mleap.FLAVOR_NAME,
-    )
-    np.testing.assert_array_almost_equal(
-        spark_model_iris.predictions, np.array(json.loads(scoring_response_2.content)), decimal=4
+        spark_model_iris.predictions, np.array(json.loads(scoring_response.content)), decimal=4
     )
 
 
@@ -592,31 +576,6 @@ def test_default_conda_env_strips_dev_suffix_from_pyspark_version(spark_model_ir
 
 
 @pytest.mark.large
-def test_mleap_model_log(spark_model_iris):
-    artifact_path = "model"
-    register_model_patch = mock.patch("mlflow.register_model")
-    with mlflow.start_run(), register_model_patch:
-        sparkm.log_model(
-            spark_model=spark_model_iris.model,
-            sample_input=spark_model_iris.spark_df,
-            artifact_path=artifact_path,
-            registered_model_name="Model1",
-        )
-        model_uri = "runs:/{run_id}/{artifact_path}".format(
-            run_id=mlflow.active_run().info.run_id, artifact_path=artifact_path
-        )
-        mlflow.register_model.assert_called_once_with(
-            model_uri, "Model1", await_registration_for=DEFAULT_AWAIT_MAX_SLEEP_SECONDS
-        )
-
-    model_path = _download_artifact_from_uri(artifact_uri=model_uri)
-    config_path = os.path.join(model_path, "MLmodel")
-    mlflow_model = Model.load(config_path)
-    assert sparkm.FLAVOR_NAME in mlflow_model.flavors
-    assert mleap.FLAVOR_NAME in mlflow_model.flavors
-
-
-@pytest.mark.large
 def test_spark_module_model_save_with_mleap_and_unsupported_transformer_raises_exception(
     spark_model_iris, model_path
 ):
@@ -634,68 +593,6 @@ def test_spark_module_model_save_with_mleap_and_unsupported_transformer_raises_e
 
 
 @pytest.mark.large
-def test_spark_module_model_save_with_relative_path_and_valid_sample_input_produces_mleap_flavor(
-    spark_model_iris,
-):
-    with TempDir(chdr=True) as tmp:
-        model_path = os.path.basename(tmp.path("model"))
-        mlflow_model = Model()
-        sparkm.save_model(
-            spark_model=spark_model_iris.model,
-            path=model_path,
-            sample_input=spark_model_iris.spark_df,
-            mlflow_model=mlflow_model,
-        )
-        assert mleap.FLAVOR_NAME in mlflow_model.flavors
-
-        config_path = os.path.join(model_path, "MLmodel")
-        assert os.path.exists(config_path)
-        config = Model.load(config_path)
-        assert mleap.FLAVOR_NAME in config.flavors
-
-
-@pytest.mark.large
-def test_mleap_module_model_save_with_relative_path_and_valid_sample_input_produces_mleap_flavor(
-    spark_model_iris,
-):
-    with TempDir(chdr=True) as tmp:
-        model_path = os.path.basename(tmp.path("model"))
-        mlflow_model = Model()
-        mleap.save_model(
-            spark_model=spark_model_iris.model,
-            path=model_path,
-            sample_input=spark_model_iris.spark_df,
-            mlflow_model=mlflow_model,
-        )
-        assert mleap.FLAVOR_NAME in mlflow_model.flavors
-
-        config_path = os.path.join(model_path, "MLmodel")
-        assert os.path.exists(config_path)
-        config = Model.load(config_path)
-        assert mleap.FLAVOR_NAME in config.flavors
-
-
-@pytest.mark.large
-def test_mleap_module_model_save_with_absolute_path_and_valid_sample_input_produces_mleap_flavor(
-    spark_model_iris, model_path
-):
-    model_path = os.path.abspath(model_path)
-    mlflow_model = Model()
-    mleap.save_model(
-        spark_model=spark_model_iris.model,
-        path=model_path,
-        sample_input=spark_model_iris.spark_df,
-        mlflow_model=mlflow_model,
-    )
-    assert mleap.FLAVOR_NAME in mlflow_model.flavors
-
-    config_path = os.path.join(model_path, "MLmodel")
-    assert os.path.exists(config_path)
-    config = Model.load(config_path)
-    assert mleap.FLAVOR_NAME in config.flavors
-
-
-@pytest.mark.large
 def test_mleap_module_model_save_with_invalid_sample_input_type_raises_exception(
     spark_model_iris, model_path
 ):
@@ -703,23 +600,6 @@ def test_mleap_module_model_save_with_invalid_sample_input_type_raises_exception
         invalid_input = pd.DataFrame()
         sparkm.save_model(
             spark_model=spark_model_iris.model, path=model_path, sample_input=invalid_input
-        )
-
-
-@pytest.mark.large
-def test_mleap_module_model_save_with_unsupported_transformer_raises_serialization_exception(
-    spark_model_iris, model_path
-):
-    class CustomTransformer(JavaModel):
-        def _transform(self, dataset):
-            return dataset
-
-    unsupported_pipeline = Pipeline(stages=[CustomTransformer()])
-    unsupported_model = unsupported_pipeline.fit(spark_model_iris.spark_df)
-
-    with pytest.raises(mleap.MLeapSerializationException):
-        mleap.save_model(
-            spark_model=unsupported_model, path=model_path, sample_input=spark_model_iris.spark_df
         )
